fix(typechecker): pin baseline vue deep imports
